refactor(playground): split code analyzer module
