+++ packages/agent-playground/src/lib/playground/code-analyzer/index.ts
@@ -1,0 +1,4 @@
+export { parseAgentConfig } from './agent-config-parser';
+export { analyzeCode } from './analysis';
+export { validateEnvironment } from './environment';
+export type { IAnalyzeCodeResult, IParsedAgentConfig, IParsedToolConfig } from './types';
